fix(wayland): read and flush compositor events on a timer (#9720)
